Added support for both .NET 8 and .NET 4.6.1
