+++ patcher/.gitignore
@@ -2,6 +2,7 @@
 bin
 obj
 /Publish
+/Res
 /HitmanPatcher.sln.DotSettings.user
 /.idea/.idea.HitmanPatcher/.idea/riderMarkupCache.xml
 *.suo
