+++ patcher/HitmanPatcher.CLI/HitmanPatcher.CLI.csproj
@@ -7,7 +7,9 @@
 
         <RootNamespace>HitmanPatcher</RootNamespace>
         <AssemblyName>PeacockPatcher.CLI</AssemblyName>
-        
+
+        <Configurations>Debug;Release;Debug - Linux;Release - Linux</Configurations>
+
         <Title>Peacock Patcher</Title>
         <Description>Peacock's HITMAN™ World of Assassination trilogy game patcher.</Description>
         <Company>The Peacock Project</Company>
@@ -18,8 +20,10 @@
         <ApplicationIcon>patcher.ico</ApplicationIcon>
         <AssemblyVersion>8.0.0.0</AssemblyVersion>
         <FileVersion>8.0.0.0</FileVersion>
+    </PropertyGroup>
 
-        <DefineConstants Condition="'$(IsLinux)' == 'true'">$(DefineConstants);LINUX</DefineConstants>
+    <PropertyGroup Condition="'$(Configuration)' == 'Release - Linux' OR '$(Configuration)' == 'Debug - Linux'">
+        <DefineConstants>$(DefineConstants);LINUX</DefineConstants>
     </PropertyGroup>
 
     <ItemGroup>
