+++ patcher/HitmanPatcher.sln
@@ -3,28 +3,40 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34316.72
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HitmanPatcher.Core", "HitmanPatcher.Core\HitmanPatcher.Core.csproj", "{70EA1386-6A9D-47C2-ADE8-2F32CDD9E946}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HitmanPatcher.Core", "HitmanPatcher.Core\HitmanPatcher.Core.csproj", "{70EA1386-6A9D-47C2-ADE8-2F32CDD9E946}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HitmanPatcher.CLI", "HitmanPatcher.CLI\HitmanPatcher.CLI.csproj", "{799814F1-919B-4BCA-9348-4F7F5F429CB9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HitmanPatcher.CLI", "HitmanPatcher.CLI\HitmanPatcher.CLI.csproj", "{799814F1-919B-4BCA-9348-4F7F5F429CB9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HitmanPatcher.UI", "HitmanPatcher.UI\HitmanPatcher.UI.csproj", "{50B900C8-B244-4AAA-9130-5C0A6093B22D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "HitmanPatcher.UI", "HitmanPatcher.UI\HitmanPatcher.UI.csproj", "{50B900C8-B244-4AAA-9130-5C0A6093B22D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug - Linux|Any CPU = Debug - Linux|Any CPU
 		Debug|Any CPU = Debug|Any CPU
+		Release - Linux|Any CPU = Release - Linux|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{70EA1386-6A9D-47C2-ADE8-2F32CDD9E946}.Debug - Linux|Any CPU.ActiveCfg = Debug - Linux|Any CPU
+		{70EA1386-6A9D-47C2-ADE8-2F32CDD9E946}.Debug - Linux|Any CPU.Build.0 = Debug - Linux|Any CPU
 		{70EA1386-6A9D-47C2-ADE8-2F32CDD9E946}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{70EA1386-6A9D-47C2-ADE8-2F32CDD9E946}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70EA1386-6A9D-47C2-ADE8-2F32CDD9E946}.Release - Linux|Any CPU.ActiveCfg = Release - Linux|Any CPU
+		{70EA1386-6A9D-47C2-ADE8-2F32CDD9E946}.Release - Linux|Any CPU.Build.0 = Release - Linux|Any CPU
 		{70EA1386-6A9D-47C2-ADE8-2F32CDD9E946}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{70EA1386-6A9D-47C2-ADE8-2F32CDD9E946}.Release|Any CPU.Build.0 = Release|Any CPU
+		{799814F1-919B-4BCA-9348-4F7F5F429CB9}.Debug - Linux|Any CPU.ActiveCfg = Debug - Linux|Any CPU
+		{799814F1-919B-4BCA-9348-4F7F5F429CB9}.Debug - Linux|Any CPU.Build.0 = Debug - Linux|Any CPU
 		{799814F1-919B-4BCA-9348-4F7F5F429CB9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{799814F1-919B-4BCA-9348-4F7F5F429CB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{799814F1-919B-4BCA-9348-4F7F5F429CB9}.Release - Linux|Any CPU.ActiveCfg = Release - Linux|Any CPU
+		{799814F1-919B-4BCA-9348-4F7F5F429CB9}.Release - Linux|Any CPU.Build.0 = Release - Linux|Any CPU
 		{799814F1-919B-4BCA-9348-4F7F5F429CB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{799814F1-919B-4BCA-9348-4F7F5F429CB9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{50B900C8-B244-4AAA-9130-5C0A6093B22D}.Debug - Linux|Any CPU.ActiveCfg = Debug|Any CPU
 		{50B900C8-B244-4AAA-9130-5C0A6093B22D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{50B900C8-B244-4AAA-9130-5C0A6093B22D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{50B900C8-B244-4AAA-9130-5C0A6093B22D}.Release - Linux|Any CPU.ActiveCfg = Release|Any CPU
 		{50B900C8-B244-4AAA-9130-5C0A6093B22D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{50B900C8-B244-4AAA-9130-5C0A6093B22D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
